feat(portal): redirect root to store and move portal
